Huh. People *use* private-use characters.
On StackOverflow, for example, some of Apple's private-use characters
are used to illustrate Mac OS keyboard commands. It would be erroneous
to consider these characters to be in need of correction.

Therefore, the "unassigned" penalty now only applies to characters that
are really unassigned (class Cn), not simply in a private use block
(class Co).