Fix logic that finds segments of text to decode and join together.
Fixes #6, by making it utterly predictable what will happen to a sequence
of greater than 2**16 characters. It will go through all the fixes
except the one that figures out if it's in the wrong encoding, because
that's the complicated one that needs more than linear time.

This is also a fix to bugs that hadn't been discovered yet, in that it
would simply break up text into lines at the wrong locations.