Fix an actual false positive.
The text "Ôôô", which appears to be an interjection that Brazilian Portuguese
speakers sometimes use online, appeared to be the MacRoman mis-encoding
of a private use character.

The heuristics now count private use and unassigned characters (that is,
unassigned in Unicode 6.1; sorry, Turkish lira) as adding 1 weirdness.
They will still decode, but they will not decode on their own when the
rare encoding penalty also applies.

Note that this is an update to char_classes.dat.