Correct some comments about the byte 0xed.
The comments here assumed, erroneously, that every instance of the byte
0xed in UTF-8 indicates that a UTF-16 surrogate is being encoded. This
isn't true; there are about 2000 Hangul characters that are encoded with
the byte 0xed.

However, the code did not rely on this assumption. It appears to decode
these Hangul characters correctly anyway, if a bit more slowly. (If it
didn't work, Twitter tests probably would have noticed by now.)