move optional imports to allow doctests to work
The streamtester modules need to import things from the 'twitter'
package, but these should not be run as part of normal testing. Doctests
were failing because they tried to import twitter, so I've moved the
import.

This had already been done in the main twitter tester, but I neglected
to also hide the import that's in "oauth.py".