Write an example differently because of Read The Docs
Yuck. Read The Docs runs on Python 2, and it interacts badly with
docstrings when "from __future__ import unicode_literals" is on.

It was already hard to document things *about* Unicode without using
Python 3 syntax. But when I use the Python 3 syntax, RTD makes it say
the wrong thing.

I'm working around this for now, but I really hope that Read The Docs
offers a Python 3 option sometime.