Update to Unicode 7
The data was built using Python 3.5a2, so that the heuristics can take
new characters in Unicode 7 into account, even when running on a version
of Python that doesn't know what they are.

3.5 is the first version to support Unicode 7, which can already be
found in some operating systems, and I'm assuming that Python's Unicode
table is not going to change between now and release.