sloppy-utf-8 was a bad approach
We have to be more conservative about where we allow utf-8 and
windows-1252 to mix. The sloppy-utf-8 approach was generating a fair
number of false positives, including quite thoroughly reinterpreting
someone's not-really-ASCII ASCII art.

It will probably work better to think of the problem as "windows-1252
with occasional UTF-8 mixed in", instead of "UTF-8 with occasional
windows-1252 mixed in".