Simplify ENDING_PUNCT_RE
It turns out that we do want to give a bonus to every instance of ending
punctuation. When two of them are next to each other, that's twice as
punctuationy and it should count. This is what fixes the Portuguese
Hogwarts false positive.

I made a change adding up the lengths of the sequences, but indeed, just
removing the + makes each character into a separate match.