revert a left-over change from the 'sloppy-utf-8' approach
At one point, in the utf-8-variants decoder, I stopped passing on
the 'final' argument and simply overrode it to True. I think the
point of this was to avoid an infinite loop when I was trying to
reuse the decoder in something called 'sloppy-utf-8', but this
turned out to just be a bad idea in many ways.

I really should be passing on final here, in case the input is
broken in the middle of a standard UTF-8 sequence.