The test case about a false positive wasn't a real false positive
Just to be clear, this has no real effect because the test case is
disabled anyway. But the thing the test case warned would happen didn't
happen on that text, and we need some slightly less probable text to
make it happen.

It would have an effect if disabled test cases said what they currently
turned into, as Andrew recommended.