Don't be quite so afraid of changing quotation marks
Fixing two-character mojibake sequences is a careful balance between
false positives and false negatives. This balance is maintained in
badness.py by `COMMON_SYMBOL_RE`, which lowers the badness of text
containing certain symbols, and `MOJIBAKE_SYMBOL_RE`, which raises the
badness in sequences that are very likely to be mojibake.

The badness was never raised again on sequences containing quotation
marks and a few other symbols, because I was overly cautious when adding
this rule. But issue #97 shows that this prevents fixing the most
common mojibake of words containing 'ë' or 'û'.

I've added these characters to the list of characters that are probably
mojibake if they appear after Â, Ã, or Ă:

    U+201A  ‚       SINGLE LOW-9 QUOTATION MARK
    U+201E  „       DOUBLE LOW-9 QUOTATION MARK
    U+2039  ‹       SINGLE LEFT-POINTING ANGLE QUOTATION MARK
    U+2018  ‘       LEFT SINGLE QUOTATION MARK
    U+201C  “       LEFT DOUBLE QUOTATION MARK
    U+201D  ”       RIGHT DOUBLE QUOTATION MARK
    U+2022  •       BULLET
    U+203A  ›       SINGLE RIGHT-POINTING ANGLE QUOTATION MARK
    U+00AB  «       LEFT-POINTING DOUBLE ANGLE QUOTATION MARK
    U+00BB  »       RIGHT-POINTING DOUBLE ANGLE QUOTATION MARK
    U+00A9  ©       COPYRIGHT SIGN

But I'm still leaving out these, which are too much risk for too little
gain:

    U+2019  ’       RIGHT SINGLE QUOTATION MARK
    U+00AE  ®       REGISTERED SIGN

Some characters have also been moved into a rule that insists on a
word-like character (`\w`) afterward, to make extra sure it's not
messing with normal punctuation.

See the new test cases for some cases that fall near the boundary of
what should be fixed.