Make the partial mojibake fixer actually work on Latin-1
The rule that locally fixes obvious Windows-1252/UTF-8 mixups (even if they
are only mixed up in part of the string) had comments that implied it would
work on Latin-1 as well.

The code assumed that the Latin-1 case was automatically handled because
this fixer would run after the encoding fixer tried replacing Latin-1
with Windows-1252. However, that step only works if it can apply
consistently across the whole string, while this fixer is meant for
fixing inconsistent mojibake. So the typical outcome is that the
encoding fixer would give up and this step wouldn't run at all.

A new test case I found in the wild revealed how this wasn't working.

Now, the local fixer for inconsistent mojibake takes responsibility for
translating Latin-1 characters itself, and we can pass the test.