Cope with different Python versions better in setup.py
This provides a more helpful message when you attempt to install ftfy on
Python 2 (there used to be a message, but in recent versions it was
erroring before it got to print the message).

Because some instances of running Python 2 are now people who are trying
to run Python 3 and messed up their environment, the message now shows
details about what version of Python is running and where it came from.

This change also sets the Trove classifier that says that we support
Python 3.7.