Better detection of mojibake of low codepoints
We have a regex that's designed to say "this looks specifically like
mojibake" that's part of the "weirdness" heuristic. It was detecting
mojibake of the Latin-1 characters. We can make this work for some more
scripts that also have low-numbered codepoints: Greek, Cyrillic, Hebrew,
and Arabic.

We had a rule that was supposed to catch when Hebrew turns into
multiplication signs -- but it wasn't working, because it wasn't
separated from the previous rule by `|`. Simply fixing that rule caused
a test case to fail, so we need a larger heuristic.