remove redundant poetry use from tests
Do not invoke the executable via `poetry` in tests, as it is redundant
and introduces unnecessary dependency on `poetry` when the package
is installed via PEP517 installer or `tox`.  When using `poetry`, one
needs to explicitly execute `poetry install` anyway and this causes
the executable to be installed.  Furthermore, the common way of running
pytest via poetry is to execute:

    poetry run pytest

causing nesting in CLI tests.